requirements.txt
fastf1>=3.6.0
pandas>=2.0.0
numpy>=1.24.0
tweepy>=4.14.0
schedule>=1.2.0
pytz>=2024.1
configparser>=5.3.0
python-dotenv>=1.0.0
requests>=2.32.0
curl_cffi>=0.7.0 